Initial commit for the R interface to Self Organising Deltoids
This is the first commit of an R interface to the self-organising deltoids
dimension squeezing algorithm. This is a simple algorithm that tries to
find reasonable positions in m-dimensional space for a set of points in
n dimensions (where m is smaller than n). It's main usage is to
visualise n-dimensional data in 2 dimensions, but any dimensionality can
be choosen.

The algorithm simply takes a set of points in N-dimensions, and then
gradually squeezes out the excess dimensions using the errors in
inter-node distances to arrange the nodes in the reduced dimensional
space.

The code in this version relies on openCL and hence requires suitable
hard and software. I also have versions that use normal multi-threading,
but I have yet to port those. That's for upcoming versions.