Optional compilation of openCL dependent components
Used autoconf to implement optional compilation of the openCL dependencies.
Autoconf will now check for the presence of the openCL libraries and
headers and if these are present will compile the full set of
functions.

If openCL is not present only the openMP accelerated objects are built,
and the openCL functionality is lost.

For this to work the files to be compiled are defined in the
configure.ac file rather than in the Makevars. There are also a few
ifdefs in the R_DimSqueezer header and implementation files.