Modified configure.ac to detect openCL frameworks on Mac OSX
Modified configure.ac to work on Mac OSX and hopefully on Linux as
well. It is now able to check for frameworks as well as libraries
and will hopefully be enough for OSX and Linux.

However, I've got little idea as to how to get things to work for
Windows.