Removed the dependence on Rcpp::Modules
Removed the dependence on Rcpp::Modules and hardcoded accessor functions
to the objects. These now make use of RefClasses and appear to be safe.

It may still be possible for the user to crash a sesssion by directly
calling .Call to squeeze or squeeze_cl and providing an external pointer
of the inappropriate type. However, apart from this, the package seems
safe enough.

Still need to change the autoconfiguration to allow for lack of openMP
on MacOS, and the issue that openCL on macOS is easy to crash by
specifying larger local work group sizes.