Completed interfaces and updated manual files
Completed the R interfaces to the C++ member functions using R reference
classes and updated the manuals to reflect their usage.

Made the R functions dependent on the system such that openCL
dependent functions are only exposed when their respective C++ classes
get compiled.

Used the AC_OPENMP macro to check for the availability of openMP and to
configure the Makevars appropriately.