Changes to enable installation on Windows
Removed bugs in the documentation that caused R CMD check to fail when
running examples in the absence of openCL.

Also added back the R/functions.R file as it will not be created
automatically on Windows. Made sure that it does not install functions
related to openCL by default. This means that it must not be removed by
the cleanup script as well.